+++ @generated/graphql/index.ts
@@ -13020,10 +13020,19 @@ export type PlanoBasicoQuery = {
       titleTemplate: string | null
     } | null
   } | null
+  allContentfulPlanos: {
+    nodes: Array<{
+      galleryTitle: string | null
+      slug: string | null
+      texto: { texto: string | null } | null
+    }>
+  }
+  allContentfulSignaturePageSubtitle: {
+    nodes: Array<{ subtitle: string | null }>
+  }
   allContentfulPlanosTextoSimples: {
     nodes: Array<{ text: { text: string | null } | null }>
   }
-  allContentfulPlanos: { nodes: Array<{ galleryTitle: string | null }> }
 }
 
 export type PlanoComunicacaoTecnologiaQueryVariables = Exact<{
@@ -13038,10 +13047,19 @@ export type PlanoComunicacaoTecnologiaQuery = {
       titleTemplate: string | null
     } | null
   } | null
+  allContentfulPlanos: {
+    nodes: Array<{
+      galleryTitle: string | null
+      slug: string | null
+      texto: { texto: string | null } | null
+    }>
+  }
+  allContentfulSignaturePageSubtitle: {
+    nodes: Array<{ subtitle: string | null }>
+  }
   allContentfulPlanosTextoSimples: {
     nodes: Array<{ text: { text: string | null } | null }>
   }
-  allContentfulPlanos: { nodes: Array<{ galleryTitle: string | null }> }
 }
 
 export type PlanoEspecialQueryVariables = Exact<{ [key: string]: never }>
@@ -13054,10 +13072,19 @@ export type PlanoEspecialQuery = {
       titleTemplate: string | null
     } | null
   } | null
+  allContentfulPlanos: {
+    nodes: Array<{
+      galleryTitle: string | null
+      slug: string | null
+      texto: { texto: string | null } | null
+    }>
+  }
+  allContentfulSignaturePageSubtitle: {
+    nodes: Array<{ subtitle: string | null }>
+  }
   allContentfulPlanosTextoSimples: {
     nodes: Array<{ text: { text: string | null } | null }>
   }
-  allContentfulPlanos: { nodes: Array<{ galleryTitle: string | null }> }
 }
 
 export type PlanoNegocioQueryVariables = Exact<{ [key: string]: never }>
@@ -13070,10 +13097,19 @@ export type PlanoNegocioQuery = {
       titleTemplate: string | null
     } | null
   } | null
+  allContentfulPlanos: {
+    nodes: Array<{
+      galleryTitle: string | null
+      slug: string | null
+      texto: { texto: string | null } | null
+    }>
+  }
+  allContentfulSignaturePageSubtitle: {
+    nodes: Array<{ subtitle: string | null }>
+  }
   allContentfulPlanosTextoSimples: {
     nodes: Array<{ text: { text: string | null } | null }>
   }
-  allContentfulPlanos: { nodes: Array<{ galleryTitle: string | null }> }
 }
 
 export type PlanoSaudeQueryVariables = Exact<{ [key: string]: never }>
@@ -13086,10 +13122,19 @@ export type PlanoSaudeQuery = {
       titleTemplate: string | null
     } | null
   } | null
+  allContentfulPlanos: {
+    nodes: Array<{
+      galleryTitle: string | null
+      slug: string | null
+      texto: { texto: string | null } | null
+    }>
+  }
+  allContentfulSignaturePageSubtitle: {
+    nodes: Array<{ subtitle: string | null }>
+  }
   allContentfulPlanosTextoSimples: {
     nodes: Array<{ text: { text: string | null } | null }>
   }
-  allContentfulPlanos: { nodes: Array<{ galleryTitle: string | null }> }
 }
 
 export type PlanoTecnologiaQueryVariables = Exact<{ [key: string]: never }>
@@ -13102,10 +13147,19 @@ export type PlanoTecnologiaQuery = {
       titleTemplate: string | null
     } | null
   } | null
+  allContentfulPlanos: {
+    nodes: Array<{
+      galleryTitle: string | null
+      slug: string | null
+      texto: { texto: string | null } | null
+    }>
+  }
+  allContentfulSignaturePageSubtitle: {
+    nodes: Array<{ subtitle: string | null }>
+  }
   allContentfulPlanosTextoSimples: {
     nodes: Array<{ text: { text: string | null } | null }>
   }
-  allContentfulPlanos: { nodes: Array<{ galleryTitle: string | null }> }
 }
 
 export type SearchPageQueryQueryVariables = Exact<{ [key: string]: never }>